Requisições HTTP para Produto e TipoProduto integrados ao frontend

+++ backend/app/repositories/product_type.py
@@ -2,6 +2,7 @@ from sqlalchemy.orm import Session
 
 from app.db.models.product_type import ProductType
 from app.utils.session_inject import with_session
+from app.db.models.product import Product
 
 
 class ProductTypeRepository:
@@ -77,9 +78,17 @@ class ProductTypeRepository:
 
     @staticmethod
     @with_session
-    def delete_product_type(product_type: ProductType, session: Session | None = None) -> None:
-        
-        existing_product_type: ProductType | None = session.get(ProductType, product_type.id)
+    def delete_product_type(id_product_type: int, session: Session | None = None) -> bool:
+        # Verifica se existe produto usando esse tipo
+        produtos = session.query(Product).filter(Product.id_tipo_produto == id_product_type).all()
+        if produtos and len(produtos) > 0:
+            # Não pode deletar, pois há produtos usando esse tipo
+            return False
+
+        existing_product_type: ProductType | None = session.get(ProductType, id_product_type)
         if existing_product_type:
             session.delete(existing_product_type)
             session.commit()
+            return True
+        else:
+            return False
